+++ animation_nodes/utils/selection.py
@@ -1,8 +1,5 @@
 import bpy
-from . handlers import eventHandler
-from . nodes import getAnimationNodeTrees
 
-enableSelectionSorting = True
 sortedSelectionNames = []
 
 def getSortedSelectedObjects():
@@ -14,31 +11,31 @@ def getSortedSelectedObjects():
 def getSortedSelectedObjectNames():
     return sortedSelectionNames
 
-@eventHandler("ALWAYS")
-def updateSelectionSorting():
-    nodeTrees = getAnimationNodeTrees()
-    if len(nodeTrees) == 0: return
-
+def updateSelectionSorting(viewLayer):
     global sortedSelectionNames
 
-    selectedNames = getSelectedObjectNames()
+    selectedNames = getSelectedObjectNames(viewLayer)
 
-    if enableSelectionSorting:
-        newSortedSelection = []
+    newSortedSelection = []
 
-        selectedNamesSet = set(selectedNames)
-        for name in sortedSelectionNames:
-            if name in selectedNamesSet:
-                newSortedSelection.append(name)
+    selectedNamesSet = set(selectedNames)
+    for name in sortedSelectionNames:
+        if name in selectedNamesSet:
+            newSortedSelection.append(name)
 
-        for name in selectedNames:
-            if name not in newSortedSelection:
-                newSortedSelection.append(name)
+    for name in selectedNames:
+        if name not in newSortedSelection:
+            newSortedSelection.append(name)
 
         sortedSelectionNames = newSortedSelection
     else:
         sortedSelectionNames = selectedNames
 
-def getSelectedObjectNames():
-    viewLayer = bpy.data.window_managers[0].windows[0].view_layer
+def getSelectedObjectNames(viewLayer):
     return [obj.name for obj in viewLayer.objects if obj.select_get(view_layer = viewLayer)]
+
+def getSelectedObjects(viewLayer):
+    return [obj for obj in viewLayer.objects if obj.select_get(view_layer = viewLayer)]
+
+def getActiveObject(viewLayer):
+    return viewLayer.objects.active