+++ CHANGELOG.md
@@ -31,6 +31,8 @@
 - Skip selection sorting if no node tree exist.
 - Vectorized *Animate Data* node.
 - Removed *Update Object Matrices* node.
+- The output of the *Get Selected Objects* node is no longer ordered.
+- Selection order now need to be recorded for *Integer ID Keys*.
 
 ## 2.2.1 (13 January 2021)